+++ backend/canisters/group/api/src/queries/events.rs
@@ -1,6 +1,6 @@
 use candid::CandidType;
 use serde::{Deserialize, Serialize};
-use types::{ChatEvent, EventIndex, EventWrapper, MessageIndex};
+use types::{EventIndex, EventsResponse, MessageIndex};
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub struct Args {
@@ -15,15 +15,8 @@ pub struct Args {
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub enum Response {
-    Success(SuccessResult),
+    Success(EventsResponse),
     CallerNotInGroup,
     ThreadMessageNotFound,
     ReplicaNotUpToDate(EventIndex),
-}
-
-#[derive(CandidType, Serialize, Deserialize, Debug)]
-pub struct SuccessResult {
-    pub events: Vec<EventWrapper<ChatEvent>>,
-    pub affected_events: Vec<EventWrapper<ChatEvent>>,
-    pub latest_event_index: EventIndex,
 }