+++ backend/canisters/group/impl/src/updates/remove_reaction.rs
@@ -37,12 +37,11 @@ fn remove_reaction_impl(args: Args, runtime_state: &mut RuntimeState) -> Respons
             thread_root_message_index: args.thread_root_message_index,
             message_id: args.message_id,
             reaction: args.reaction,
-            correlation_id: args.correlation_id,
             now,
         }) {
-            AddRemoveReactionResult::Success(r) => {
+            AddRemoveReactionResult::Success => {
                 handle_activity_notification(runtime_state);
-                SuccessV2(r)
+                Success
             }
             AddRemoveReactionResult::NoChange => NoChange,
             AddRemoveReactionResult::MessageNotFound => MessageNotFound,