+++ backend/canisters/proposals_bot/impl/src/lifecycle/inspect_message.rs
@@ -10,7 +10,9 @@ fn accept_if_valid(runtime_state: &RuntimeState) {
     let method_name = ic_cdk::api::call::method_name();
 
     let is_valid = match method_name.as_str() {
-        "add_governance_canister" | "remove_governance_canister" => runtime_state.is_caller_governance_principal(),
+        "add_governance_canister" | "remove_governance_canister" | "appoint_admins" => {
+            runtime_state.is_caller_governance_principal()
+        }
         _ => false,
     };