+++ frontend/openchat-agent/src/services/group/candid/types.d.ts
@@ -47,7 +47,7 @@ export type AddReactionResponse = { 'MessageNotFound' : null } |
   { 'CallerNotInGroup' : null } |
   { 'ChatFrozen' : null } |
   { 'NotAuthorized' : null } |
-  { 'Success' : EventIndex } |
+  { 'Success' : null } |
   { 'UserSuspended' : null } |
   { 'InvalidReaction' : null } |
   { 'SuccessV2' : PushEventResult };
@@ -184,10 +184,6 @@ export interface ChatMetrics {
   'reactions' : bigint,
   'prize_messages' : bigint,
 }
-export type ChatSummary = { 'Group' : GroupChatSummary } |
-  { 'Direct' : DirectChatSummary };
-export type ChatSummaryUpdates = { 'Group' : GroupChatSummaryUpdates } |
-  { 'Direct' : DirectChatSummaryUpdates };
 export interface ChatUnfrozen { 'unfrozen_by' : UserId }
 export interface ClaimPrizeArgs {
   'correlation_id' : bigint,
@@ -347,7 +343,6 @@ export type EventsResponse = { 'ThreadMessageNotFound' : null } |
   { 'CallerNotInGroup' : null } |
   { 'Success' : EventsSuccessResult };
 export interface EventsSuccessResult {
-  'affected_events' : Array<ChatEventWrapper>,
   'events' : Array<ChatEventWrapper>,
   'latest_event_index' : number,
 }
@@ -462,7 +457,6 @@ export interface GroupCanisterGroupChatSummaryUpdates {
   'latest_event_index' : [] | [EventIndex],
   'mentions' : Array<Mention>,
   'chat_id' : ChatId,
-  'affected_events_v2' : Array<[EventIndex, TimestampMillis]>,
   'newly_expired_messages' : Array<MessageIndexRange>,
   'participant_count' : [] | [number],
   'my_metrics' : [] | [ChatMetrics],
@@ -510,36 +504,6 @@ export interface GroupChatSummary {
   'archived' : boolean,
   'participant_count' : number,
   'my_metrics' : ChatMetrics,
-  'latest_message' : [] | [MessageEventWrapper],
-}
-export interface GroupChatSummaryUpdates {
-  'is_public' : [] | [boolean],
-  'permissions' : [] | [GroupPermissions],
-  'metrics' : [] | [ChatMetrics],
-  'subtype' : GroupSubtypeUpdate,
-  'date_last_pinned' : [] | [TimestampMillis],
-  'name' : [] | [string],
-  'role' : [] | [Role],
-  'wasm_version' : [] | [Version],
-  'affected_events' : Uint32Array | number[],
-  'notifications_muted' : [] | [boolean],
-  'description' : [] | [string],
-  'events_ttl' : [] | [Milliseconds],
-  'last_updated' : TimestampMillis,
-  'owner_id' : [] | [UserId],
-  'avatar_id' : AvatarIdUpdate,
-  'next_message_expiry' : TimestampUpdate,
-  'latest_threads' : Array<ThreadSyncDetails>,
-  'frozen' : FrozenGroupUpdate,
-  'latest_event_index' : [] | [EventIndex],
-  'read_by_me_up_to' : [] | [MessageIndex],
-  'mentions' : Array<Mention>,
-  'chat_id' : ChatId,
-  'date_read_pinned' : [] | [TimestampMillis],
-  'expired_messages' : Array<MessageIndexRange>,
-  'archived' : [] | [boolean],
-  'participant_count' : [] | [number],
-  'my_metrics' : [] | [ChatMetrics],
   'latest_message' : [] | [MessageEventWrapper],
 }
 export interface GroupDescriptionChanged {
@@ -1016,7 +980,7 @@ export type RemoveReactionResponse = { 'MessageNotFound' : null } |
   { 'CallerNotInGroup' : null } |
   { 'ChatFrozen' : null } |
   { 'NotAuthorized' : null } |
-  { 'Success' : EventIndex } |
+  { 'Success' : null } |
   { 'UserSuspended' : null } |
   { 'SuccessV2' : PushEventResult };
 export interface ReplyContext {