+++ frontend/openchat-shared/src/domain/chat/chat.ts
@@ -683,7 +683,6 @@ export type EventWrapper<T extends ChatEvent> = {
 
 export type EventsSuccessResult<T extends ChatEvent> = {
     events: EventWrapper<T>[];
-    affectedEvents: EventWrapper<T>[];
     latestEventIndex: number | undefined;
 };